# openapi-docs.yaml
openapi: 3.0.0
info:
  description: |
    This is the OpenAPI documentation for the TakeNote REST API. You can find
    out more at [https://app.swaggerhub.com/apis/vladpirlog/takenote-express/1.0.0]
    (https://app.swaggerhub.com/apis/vladpirlog/takenote-express/1.0.0).
  version: "1.0.0"
  title: TakeNote REST API
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: avi.pirlog@gmail.com
  license:
    name: MIT License
    url: "https://opensource.org/licenses/MIT"
servers:
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/vladpirlog/takenote-express/1.0.0
  - description: When running locally on port 8000
    url: "http://localhost:8000/"
...
paths:
...
  "/notes/{noteID}/tags":
    post:
      tags:
        - notes
      summary: Add tags
      operationId: addTags
      security:
        - api_key: []
      parameters:
        - name: noteID
          in: path
          description: ID of a note to add tags to
          required: true
          schema:
            $ref: "#/components/schemas/Note/properties/_id"
        - name: tags
          in: query
          description: Tags to add to the note, separated with commas
          schema:
            type: string
      responses:
        "401":
          description: No authenticated user
        "403":
          description: User has incompatible state
        "422":
          description: Input data missing from path or query
        "400":
          description: Couldn't add tags or tags invalid
        "200":
          description: Tags added
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    format: int32
                    description: HTTP status code
                  message:
                    type: string
                  tags:
                    $ref: "#/components/schemas/Note/properties/tags"
    delete:
      tags:
        - notes
      summary: Delete tags
      operationId: deleteTags
      security:
        - api_key: []
      parameters:
        - name: noteID
          in: path
          description: ID of a note to delete tags from
          required: true
          schema:
            $ref: "#/components/schemas/Note/properties/_id"
        - name: tags
          in: query
          description: Tags to delete from the note, separated with commas
          schema:
            type: string
      responses:
        "401":
          description: No authenticated user
        "403":
          description: User has incompatible state
        "422":
          description: Input data missing from path or query
        "400":
          description: Couldn't delete tags or tags invalid
        "200":
          description: Tags deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    format: int32
                    description: HTTP status code
                  message:
                    type: string
...
components:
...
  securitySchemes:
    api_key:
      type: apiKey
      description: JWT signed token used for authentication
      name: api_key
      in: cookie